Optional Digestion Count Output (#2460)
* Add WriteDigestCountCheckBox to SearchTaskWindow and Search Task Parameters

* Add DictionaryExtensions class with AddOrCreate method with tests

* moved anlyte type to util folder

* Add Increment method to DictionaryExtensions

Added a generic Increment method to DictionaryExtensions.cs that increments the value of a specified key or initializes it to one if the key does not exist. Included XML documentation for the method. Updated DictionaryExtensionsTests.cs with unit tests covering various scenarios for the Increment method.

* Add feature to track digestion product counts per protein to ClassicSearchEngine

* Only write tsv file for classic search

* Add IsNullOrEmpty method to DictionaryExtensions

Added IsNullOrEmpty method to DictionaryExtensions.cs to check if a dictionary is null or empty. Included XML documentation for the method. Added unit tests in DictionaryExtensionsTests.cs to verify the method's behavior for null, empty, and non-empty dictionaries.

* Add feature to track and write digestion product counts

Introduced functionality to track and write digestion product counts for proteins during a search task. Key changes include:

- Made `DigestionCountDictionary` a public readonly field in `ClassicSearchEngine.cs` and adjusted the constructor accordingly.
- Added an internal property `DigestionCountDictionary` in `PostSearchAnalysisTask.cs` and implemented methods to write counts to .tsv files.
- Modified `SearchTask.cs` to initialize and pass `digestionCountDictionary` to `PostSearchAnalysisTask`.
- Added tests in `PostSearchAnalysisTaskTests.cs` to verify the correct writing of digestion counts and histograms.

* Ensured writing consistency with multiple data files

* Cleanup and Additional Comments

* Added more comments to tooltop

* Ensured thread safety for extension methods with concurent dictionary.

* Added tests for thread safety

* AddOrCreateThreadSafe take in lock object

* Adjusted Dictionary Extensions Increment method to

* Update digestion count tracking and file output format

Updated `DigestionCountDictionary` to track by protein accession and base sequence. Modified `PostSearchAnalysisTask` and `SearchTask` to use the new type. Updated file output logic to include primary sequence and added checks for `WriteDecoys` parameter. Enhanced unit tests to reflect these changes and added new tests for decoy handling.