Meta draw settings to singleton (#2529)
* Broke out settings view in MetaDraw to have its own control for dropping into other controls

* Refactor MetaDrawSettingsWindow to use singleton ViewModel

Removed the constructor dependency on `MetaDrawSettingsViewModel` and replaced it with a direct reference to a singleton instance. Updated the `PopulateChoices`, `Save`, and `SaveAsDefault` methods to utilize the singleton instance, ensuring consistent data handling. Adjusted the settings reset logic to load from the singleton instance. Modified `SettingsButtonControl` to instantiate `MetaDrawSettingsWindow` without passing a `SettingsViewModel`.

* New Test

* Made constructor internal, bound and tested some public properties

* Update xaml.cs

* Reduced max height of settings window

---------

Co-authored-by: trishorts <mshort@chem.wisc.edu>